Feature - After Action Reports Screenshot Archives
What
=================
Created self monitoring screenshot archives based on images files in
directory as well as a "description.txt" file for any descriptions to
describe something about that particular operation.

Why
=================
To provide a place for unit Screenshots to be stored. Needs to be moved
somewhere more maintainable in the future.